+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Redesign normalize path to ctest
 - Removed `WsjcppCore::doNormalizePath`
 - Fixed code-formatting
+- Renamed `WsjcppCore::extractURLProtocol` to `wsjcpp::extract_url_protocol`
 
 ## [v0.2.5] - 2026-05-19 (2026 May 19)